Add role to user api

+++ src/modules/specialty/specialty.service.ts
@@ -1,16 +1,20 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Not, Repository } from 'typeorm';
 import { Specialty } from './entities/specialty.entity';
 import { CreateSpecialtyDto } from './dto/createSpecialty.dto';
 import { ERROR_MESSAGES } from 'src/common/constants/errorMessage.constant';
 import { BadRequestException } from 'src/common/exceptions/badRequest.exception';
+import { Doctor } from '../doctors/entities/doctor.entity';
 
 @Injectable()
 export class SpecialtyService {
   constructor(
     @InjectRepository(Specialty)
-    private readonly repo: Repository<Specialty>
+    private readonly repo: Repository<Specialty>,
+
+    @InjectRepository(Doctor)
+    private readonly doctorRepo: Repository<Doctor>
   ) {}
 
   async findAll() {
@@ -57,17 +61,24 @@ export class SpecialtyService {
   }
 
   async update(id: number, dto: CreateSpecialtyDto) {
+    //  1. Kiểm tra trùng tên với chuyên khoa khác
     const existName = await this.repo.findOne({
-      where: { name: dto.name, id: Not(id) }
+      where: {
+        name: dto.name,
+        id: Not(id)
+      }
     });
 
-    if (existName)
-      throw new NotFoundException({
-        message: ERROR_MESSAGES.specialty.SPECIALTY_NOT_FOUND
+    if (existName) {
+      throw new BadRequestException({
+        message: ERROR_MESSAGES.specialty.SPECIALTY_NAME_ALREADY_EXISTS
       });
+    }
 
+    //  2. Kiểm tra chuyên khoa có tồn tại không
     const exist = await this.findOne(id);
 
+    //  3. Merge & save
     const merged = { ...exist.data, ...dto };
     const data = await this.repo.save(merged);
 
@@ -79,6 +90,19 @@ export class SpecialtyService {
 
   async remove(id: number) {
     const exist = await this.findOne(id);
+
+    const doctorCount = await this.doctorRepo.count({
+      where: {
+        specialty: { id }
+      }
+    });
+
+    if (doctorCount > 0) {
+      throw new ConflictException({
+        message: 'SPECIALTY_IS_IN_USE',
+        detail: `Chuyên khoa đang có ${doctorCount} bác sĩ, không thể xoá`
+      });
+    }
 
     await this.repo.remove(exist.data);
